+++ src/divider/sectionDivider/index.tsx
@@ -1,0 +1,1 @@
+export { SectionDivider } from './SectionDivider'